+++ noweda/__init__.py
@@ -1,4 +1,6 @@
 from .io import read, read_chunked
 from .accessor import NowEDAAccessor
 
-__all__ = ["read", "read_chunked"]
+from ._version import __version__
+
+__all__ = ["read", "read_chunked", "__version__"]